[ADD] listObjet Ajout d'un fichier pour mettre la liste des éléments à afficher
[UPDATE] Vector
Cast de la valeur de retour de sqrt en float

[UPDATE] Scene
Ajout du constructeur / destructeur

[UPDATE] Main
Ajout d'un main rapide

[UPDATE] Cylinder
Cast de la valeur de retour de sqrt en float

+++ src/Vector.cpp
@@ -31,7 +31,7 @@ float Vector::z() const {
 }
 
 float Vector::norm() const {
-    return sqrt(m_x * m_x + m_y * m_y + m_z * m_z);
+    return (float) sqrt(m_x * m_x + m_y * m_y + m_z * m_z);
 }
 
 float Vector::produitScalaire(Vector const &o) const {
